Move post specific dropdown items into their own directory
