+++ packages/lesswrong/components/dropdowns/posts/MarkAsReadDropdownItem.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { registerComponent, Components } from "../../lib/vulcan-lib";
-import { preferredHeadingCase } from "../../lib/forumTypeUtils";
-import { useItemsRead } from "../hooks/useRecordPostView";
-import { useNamedMutation } from "../../lib/crud/withMutation";
+import { registerComponent, Components } from "../../../lib/vulcan-lib";
+import { preferredHeadingCase } from "../../../lib/forumTypeUtils";
+import { useItemsRead } from "../../hooks/useRecordPostView";
+import { useNamedMutation } from "../../../lib/crud/withMutation";
 
 const MarkAsReadDropdownItem = ({post}: {post: PostsBase}) => {
   const {postsRead, setPostRead} = useItemsRead();
