+++ packages/lesswrong/components/dropdowns/posts/ShortformDropdownItem.tsx
@@ -1,9 +1,9 @@
 import React from "react";
-import { registerComponent, Components } from "../../lib/vulcan-lib";
-import { preferredHeadingCase } from "../../lib/forumTypeUtils";
-import { useCurrentUser } from "../common/withUser";
-import { useUpdate } from "../../lib/crud/withUpdate";
-import { userCanDo } from "../../lib/vulcan-users";
+import { registerComponent, Components } from "../../../lib/vulcan-lib";
+import { preferredHeadingCase } from "../../../lib/forumTypeUtils";
+import { useCurrentUser } from "../../common/withUser";
+import { useUpdate } from "../../../lib/crud/withUpdate";
+import { userCanDo } from "../../../lib/vulcan-users";
 
 const ShortformDropdownItem = ({post}: {post: PostsBase}) => {
   const currentUser = useCurrentUser();
