+++ packages/lesswrong/lib/components.ts
@@ -82,24 +82,24 @@ importComponent("DraftJSEditor", () => require('../components/editor/DraftJSEdit
 // Dropdowns
 importComponent("DropdownMenu", () => require('../components/dropdowns/DropdownMenu'));
 importComponent("DropdownItem", () => require('../components/dropdowns/DropdownItem'));
-importComponent("SuggestCuratedDropdownItem", () => require('../components/dropdowns/SuggestCuratedDropdownItem'));
-importComponent("DeleteDraftDropdownItem", () => require('../components/dropdowns/DeleteDraftDropdownItem'));
-importComponent("MoveToDraftDropdownItem", () => require('../components/dropdowns/MoveToDraftDropdownItem'));
-importComponent("MarkAsReadDropdownItem", () => require('../components/dropdowns/MarkAsReadDropdownItem'));
-importComponent("SummarizeDropdownItem", () => require('../components/dropdowns/SummarizeDropdownItem'));
-importComponent("MoveToFrontpageDropdownItem", () => require('../components/dropdowns/MoveToFrontpageDropdownItem'));
-importComponent("SuggestAlignmentPostDropdownItem", () => require('../components/dropdowns/SuggestAlignmentPostDropdownItem'));
-importComponent("MoveToAlignmentDropdownItem", () => require('../components/dropdowns/MoveToAlignmentDropdownItem'));
-importComponent("ShortformDropdownItem", () => require('../components/dropdowns/ShortformDropdownItem'));
-importComponent("ApproveNewUserDropdownItem", () => require('../components/dropdowns/ApproveNewUserDropdownItem'));
-importComponent("EditTagsDropdownItem", () => require('../components/dropdowns/EditTagsDropdownItem'));
-importComponent("ReportPostDropdownItem", () => require('../components/dropdowns/ReportPostDropdownItem'));
-importComponent("HideFrontpagePostDropdownItem", () => require('../components/dropdowns/HideFrontpagePostDropdownItem'));
-importComponent("BookmarkDropdownItem", () => require('../components/dropdowns/BookmarkDropdownItem'));
-importComponent("EditPostDropdownItem", () => require('../components/dropdowns/EditPostDropdownItem'));
-importComponent("DuplicateEventDropdownItem", () => require('../components/dropdowns/DuplicateEventDropdownItem'));
-importComponent("PostAnalyticsDropdownItem", () => require('../components/dropdowns/PostAnalyticsDropdownItem'));
 importComponent("NotifyMeDropdownItem", () => require('../components/dropdowns/NotifyMeDropdownItem'));
+importComponent("SuggestCuratedDropdownItem", () => require('../components/dropdowns/posts/SuggestCuratedDropdownItem'));
+importComponent("DeleteDraftDropdownItem", () => require('../components/dropdowns/posts/DeleteDraftDropdownItem'));
+importComponent("MoveToDraftDropdownItem", () => require('../components/dropdowns/posts/MoveToDraftDropdownItem'));
+importComponent("MarkAsReadDropdownItem", () => require('../components/dropdowns/posts/MarkAsReadDropdownItem'));
+importComponent("SummarizeDropdownItem", () => require('../components/dropdowns/posts/SummarizeDropdownItem'));
+importComponent("MoveToFrontpageDropdownItem", () => require('../components/dropdowns/posts/MoveToFrontpageDropdownItem'));
+importComponent("SuggestAlignmentPostDropdownItem", () => require('../components/dropdowns/posts/SuggestAlignmentPostDropdownItem'));
+importComponent("MoveToAlignmentDropdownItem", () => require('../components/dropdowns/posts/MoveToAlignmentDropdownItem'));
+importComponent("ShortformDropdownItem", () => require('../components/dropdowns/posts/ShortformDropdownItem'));
+importComponent("ApproveNewUserDropdownItem", () => require('../components/dropdowns/posts/ApproveNewUserDropdownItem'));
+importComponent("EditTagsDropdownItem", () => require('../components/dropdowns/posts/EditTagsDropdownItem'));
+importComponent("ReportPostDropdownItem", () => require('../components/dropdowns/posts/ReportPostDropdownItem'));
+importComponent("HideFrontpagePostDropdownItem", () => require('../components/dropdowns/posts/HideFrontpagePostDropdownItem'));
+importComponent("BookmarkDropdownItem", () => require('../components/dropdowns/posts/BookmarkDropdownItem'));
+importComponent("EditPostDropdownItem", () => require('../components/dropdowns/posts/EditPostDropdownItem'));
+importComponent("DuplicateEventDropdownItem", () => require('../components/dropdowns/posts/DuplicateEventDropdownItem'));
+importComponent("PostAnalyticsDropdownItem", () => require('../components/dropdowns/posts/PostAnalyticsDropdownItem'));
 
 // RSS Feed Integration
 importComponent("newFeedButton", () => require('../components/feeds/newFeedButton'));
